fix: match apidoc for `ModelSingle` to actual return type
While this endpoint *should* return a single value, in reality it does (and probably always has) actually return a list of a single element instead. Because changing the actual return value constitutes a breaking change, I recommend simply fixing the API doc for now and changing it to the (presumably) intended type when we go to `/v3/`.