fix: don't iterate on `None` from gen_metadata key
`dict.get(...)` returns the value of a key if it is set. The default parameter (the second parameter) only returns a default value **if the key does not exist**. Therefore, if the key `gen_metadata` is set in `kwargs`, and is set to `None`, then `kwargs.get("gen_metadata", [])` will return `None` and **not** an empty list (`[]`).