fix: allow `outpainting` a source_processing type
`outpainting` is allowed as a `source_processing` value, but prior to this change setting this value will cause a 500 stemming from a `sqlalchemy.exc.DataError: (psycopg2.errors.StringDataRightTruncation) value too long for type character varying(10)` error.

This change allows the intended number of characters, plus an extra one for posterity.